requirements.txt
pillow
bing-image-downloader
scikit-learn
numpy
opencv-python
torch
tensorflow
flask
uvicorn
gradio
requests
ultralytics
torch

